pygraphql/requirements.txt
PyYAML
ariadne>=0.16.0
websockets
uvicorn
gql
